provide markdown to plaintext conversion for mailers
